+++ src/Lykke.Job.TransactionHandler.Core/Services/Messages/Email/ContentGenerator/MessagesData/NoRefundOCashOutData.cs
@@ -3,7 +3,7 @@
     public class NoRefundOCashOutData : IEmailMessageData
     {
         public string AssetId { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
         public string SrcBlockchainHash { get; set; }
         public string MessageId()
         {
